PWHistList is now a std::vector instead of a CList - make pwsafe a wee bit more portable (part 1)
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@1357 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/EditDlg.h
@@ -47,8 +47,8 @@ private:
   time_t m_tttLTime;
   bool m_bIsModified;
   // Password History related stuff
-  int m_NumPWHistory;
-  int m_MaxPWHistory;
+  size_t m_NumPWHistory;
+  size_t m_MaxPWHistory;
   BOOL m_SavePWHistory;
   PWHistList m_PWHistList;
   CMyString m_PWHistory;
